HoH Slave/requirements.txt
aiohttp==3.8.3
discord.py==2.3.1
gspread==5.7.2
numpy==1.24.3
opencv_python_headless==4.8.0.74
Pillow==10.0.0
pytesseract==0.3.10
python-dotenv==1.0.0
ultralytics==8.0.145
